requirements.txt
###### Requirements without Version Specifiers ######
urllib3
simplejson
python-dotenv